% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{candles_module}
\alias{candles_module}
\title{Renders the Candlestick Chart Module with Trades}
\usage{
candles_module(mktdata, txns, theme)
}
\arguments{
\item{mktdata}{An xts object with OHLC market data.}

\item{txns}{An xts object with the transactions (trades).}

\item{theme}{The theme list object.}
}
\value{
A highchart object.
}
\description{
Renders the Candlestick Chart Module with Trades
}
\keyword{internal}
