% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-backtest.R
\name{costs_module}
\alias{costs_module}
\title{Renders the Costs/Friction Module}
\usage{
costs_module(costs_df, theme)
}
\arguments{
\item{costs_df}{A normalized costs data.frame.}

\item{theme}{The theme list object.}
}
\value{
An HTML/highcharter module.
}
\description{
Renders the Costs/Friction Module
}
\keyword{internal}
