% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{default_theme}
\alias{default_theme}
\title{Default Theme for tplot Charts}
\usage{
default_theme()
}
\value{
A list of theme settings.
}
\description{
Returns a list with color and font settings for the light theme.
}
