% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.as_backtest}
\alias{.as_backtest}
\title{Try to interpret a list as a backtest-like object}
\usage{
.as_backtest(obj)
}
\arguments{
\item{obj}{The list to check.}
}
\value{
A list with mktdata, trades, and stats, or NULL.
}
\description{
Try to interpret a list as a backtest-like object
}
\keyword{internal}
