% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.available_modules}
\alias{.available_modules}
\title{Determine which modules can be rendered based on prepared data}
\usage{
.available_modules(prep)
}
\arguments{
\item{prep}{The list of data prepared by .tplot_prepare.}
}
\value{
A character vector with the names of available modules.
}
\description{
Determine which modules can be rendered based on prepared data
}
\keyword{internal}
