% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.data_prepare}
\alias{.data_prepare}
\title{Prepares raw data by calculating returns}
\usage{
.data_prepare(data_xts, verbose = getOption("tplot.verbose", FALSE))
}
\arguments{
\item{data_xts}{A list of xts objects.}
}
\value{
A single xts object with the calculated returns for all series.
}
\description{
Prepares raw data by calculating returns
}
\keyword{internal}
