% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.find_series_in_env}
\alias{.find_series_in_env}
\title{Find any xts candidates for the provided names in the environment}
\usage{
.find_series_in_env(symbols, init, finit)
}
\arguments{
\item{symbols}{A character vector of symbol names.}

\item{init}{Start date.}

\item{finit}{End date.}
}
\value{
A list of found xts objects.
}
\description{
Find any xts candidates for the provided names in the environment
}
\keyword{internal}
