% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sources.R
\name{.get_object_if_exists}
\alias{.get_object_if_exists}
\title{Get an object by name from common environments}
\usage{
.get_object_if_exists(name, envs = NULL)
}
\arguments{
\item{name}{The name of the object to look for.}

\item{envs}{A list of environments to search in.}
}
\value{
The object, or NULL if not found.
}
\description{
Get an object by name from common environments
}
\keyword{internal}
