% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.is_ohlc_xts}
\alias{.is_ohlc_xts}
\title{Check if an xts object has OHLC columns}
\usage{
.is_ohlc_xts(x)
}
\arguments{
\item{x}{The xts object.}
}
\value{
TRUE if it has OHLC columns, FALSE otherwise.
}
\description{
Check if an xts object has OHLC columns
}
\keyword{internal}
