% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sources.R
\name{.is_xts}
\alias{.is_xts}
\title{Check if an object is of class xts}
\usage{
.is_xts(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
TRUE if the object is an xts object, FALSE otherwise.
}
\description{
Check if an object is of class xts
}
\keyword{internal}
