% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di-futures.R
\name{.isDI}
\alias{.isDI}
\title{Checks if market data is for brazilian DI Futures}
\usage{
.isDI(mkt)
}
\arguments{
\item{mkt}{The xts object with market data.}
}
\value{
Logical.
}
\description{
Checks if market data is for brazilian DI Futures
}
\keyword{internal}
