% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sources.R
\name{.quantstrat_portfolio_data}
\alias{.quantstrat_portfolio_data}
\title{Try to resolve a blotter/quantstrat portfolio by name}
\usage{
.quantstrat_portfolio_data(name, init = NULL, finit = NULL)
}
\arguments{
\item{name}{Portfolio name (character).}

\item{init}{Start date to subset.}

\item{finit}{End date to subset.}
}
\value{
list(mktdata=xts or NULL, trades=xts or NULL, symbol=character) or NULL
}
\description{
Try to resolve a blotter/quantstrat portfolio by name
}
\keyword{internal}
