% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.subset_xts}
\alias{.subset_xts}
\title{Safely subset an xts object by a date range}
\usage{
.subset_xts(x, init, finit)
}
\arguments{
\item{x}{The xts object.}

\item{init}{Start date.}

\item{finit}{End date.}
}
\value{
The subsetted xts object.
}
\description{
Safely subset an xts object by a date range
}
\keyword{internal}
