% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.to_ohlc_standard}
\alias{.to_ohlc_standard}
\title{Coerce an xts object to a standard OHLC shape expected by highcharter
Names the columns as Open, High, Low, Close, optionally preserving Volume.}
\usage{
.to_ohlc_standard(x)
}
\arguments{
\item{x}{xts object with any OHLC-like column names}
}
\value{
xts with 4 (or 5 with Volume) columns or NULL if cannot map
}
\description{
Coerce an xts object to a standard OHLC shape expected by highcharter
Names the columns as Open, High, Low, Close, optionally preserving Volume.
}
\keyword{internal}
