% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{.tplot_render_html}
\alias{.tplot_render_html}
\title{Renders the tplot output in HTML format}
\usage{
.tplot_render_html(
  prep,
  modules,
  theme = dark_theme(),
  output_dir = "~",
  viewer = FALSE
)
}
\arguments{
\item{prep}{The list of data prepared by .tplot_prepare.}

\item{modules}{The modules to include.}

\item{theme}{The theme list object.}

\item{output_dir}{The output directory.}

\item{viewer}{Logical, whether to open in the RStudio viewer instead of saving a file.}
}
\value{
The path to the saved HTML file.
}
\description{
Renders the tplot output in HTML format
}
\keyword{internal}
