% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{.tplot_render_json}
\alias{.tplot_render_json}
\title{Renders the tplot output in JSON format}
\usage{
.tplot_render_json(prep, modules)
}
\arguments{
\item{prep}{The list of data prepared by .tplot_prepare.}

\item{modules}{The modules to include.}
}
\value{
A JSON string.
}
\description{
Renders the tplot output in JSON format
}
\keyword{internal}
