% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.uniquify_names}
\alias{.uniquify_names}
\title{Ensure unique names in a named list by appending _2, _3, ... for duplicates}
\usage{
.uniquify_names(x)
}
\arguments{
\item{x}{a named list}
}
\value{
same list with unique names
}
\description{
Ensure unique names in a named list by appending _2, _3, ... for duplicates
}
\keyword{internal}
