% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-performance.R
\name{drawdown_module}
\alias{drawdown_module}
\title{Renders the Drawdowns Chart Module}
\usage{
drawdown_module(
  drawdowns,
  datas,
  ativo,
  benchs,
  theme,
  sync_with_candles = FALSE
)
}
\arguments{
\item{drawdowns}{An xts object with the drawdown series.}

\item{datas}{A numeric vector of dates in milliseconds.}

\item{ativo}{The name of the main asset.}

\item{benchs}{A character vector of benchmark names.}

\item{theme}{The theme list object.}

\item{sync_with_candles}{Logical, whether to sync with the candlestick chart.}
}
\value{
A highchart object.
}
\description{
Renders the Drawdowns Chart Module
}
\keyword{internal}
