% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fix_pkg}
\alias{fix_pkg}
\title{Fixes HTML dependencies with symbolic package names}
\usage{
fix_pkg(x)
}
\arguments{
\item{x}{The HTML dependency object or list.}
}
\value{
The corrected object.
}
\description{
Fixes HTML dependencies with symbolic package names
}
\keyword{internal}
