% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{footer_module}
\alias{footer_module}
\title{Renders the Footer Module}
\usage{
footer_module(theme)
}
\arguments{
\item{theme}{The theme list object.}
}
\value{
An HTML object containing the footer text.
}
\description{
Renders the Footer Module
}
\keyword{internal}
