% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di-futures.R
\name{get_DI_price}
\alias{get_DI_price}
\title{Calculates the rate from the PU of a brazilian DI futures contract}
\usage{
get_DI_price(price, row_date, maturity)
}
\arguments{
\item{price}{The PU (unit price) of the contract.}

\item{row_date}{The date of the calculation.}

\item{maturity}{The maturity date of the contract.}
}
\value{
The numeric rate.
}
\description{
Calculates the rate from the PU of a brazilian DI futures contract
}
\keyword{internal}
