% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{The value to test.}

\item{b}{The fallback value if 'a' is NULL.}
}
\value{
'a' if it is not NULL, otherwise 'b'.
}
\description{
Returns the left-hand side if it's not NULL,
otherwise returns the right-hand side.
}
\keyword{internal}
