% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{periodret_module}
\alias{periodret_module}
\title{Renders the Period Returns Chart Module}
\usage{
periodret_module(
  ret_simple,
  datas,
  ativo,
  benchs,
  theme,
  sync_with_candles = FALSE
)
}
\arguments{
\item{ret_simple}{An xts object with simple returns.}

\item{datas}{A numeric vector of dates in milliseconds.}

\item{ativo}{The name of the main asset.}

\item{benchs}{A character vector of benchmark names.}

\item{theme}{The theme list object.}

\item{sync_with_candles}{Logical, whether to sync with the candlestick chart.}
}
\value{
A highchart object.
}
\description{
Renders the Period Returns Chart Module
}
\keyword{internal}
