% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-market.R
\name{position_module}
\alias{position_module}
\title{Renders the Position Chart Module}
\usage{
position_module(mktdata, txns, theme, sync_with_candles = FALSE)
}
\arguments{
\item{mktdata}{An xts object with market data.}

\item{txns}{An xts object with transactions.}

\item{theme}{The theme list object.}

\item{sync_with_candles}{Logical, whether to sync with the candlestick chart.}
}
\value{
A highchart object.
}
\description{
Renders the Position Chart Module
}
\keyword{internal}
