% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rentab_json}
\alias{rentab_json}
\title{Converts the list of returns tables to a JSON-ready list}
\usage{
rentab_json(lista_tabs)
}
\arguments{
\item{lista_tabs}{The list of returns tables.}
}
\value{
A list ready to be converted to JSON.
}
\description{
Converts the list of returns tables to a JSON-ready list
}
\keyword{internal}
