% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-performance.R
\name{rollingcorr_module}
\alias{rollingcorr_module}
\title{Renders the Rolling Correlation Chart Module}
\usage{
rollingcorr_module(rolling_corr, datas, theme, sync_with_candles = FALSE)
}
\arguments{
\item{rolling_corr}{An xts object with rolling pairwise correlations.}

\item{datas}{A numeric vector of dates in milliseconds.}

\item{theme}{The theme list object.}

\item{sync_with_candles}{Logical, whether to sync with the candlestick chart.}
}
\value{
A highchart object.
}
\description{
Renders the Rolling Correlation Chart Module
}
\keyword{internal}
