% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-performance.R
\name{rollingret_module}
\alias{rollingret_module}
\title{Renders the Rolling Returns Chart Module}
\usage{
rollingret_module(
  ret_cum,
  datas,
  ativo,
  benchs,
  theme,
  sync_with_candles = FALSE
)
}
\arguments{
\item{ret_cum}{An xts object with cumulative returns.}

\item{datas}{A numeric vector of dates in milliseconds.}

\item{ativo}{The name of the main asset.}

\item{benchs}{A character vector of benchmark names.}

\item{theme}{The theme list object.}

\item{sync_with_candles}{Logical, whether to sync with the candlestick chart.}
}
\value{
A highchart object.
}
\description{
Renders the Rolling Returns Chart Module
}
\keyword{internal}
