% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTradingPlots-package.R
\docType{package}
\name{rTradingPlots-package}
\alias{rTradingPlots}
\alias{rTradingPlots-package}
\title{rTradingPlots: Generate Customizable Interactive Performance Reports for Financial Assets}
\description{
R library for easy, informative trading/backtesting charts in HTML (self-embed option, JPG, PNG). Custom themes. Integrates with quantstrat/blotter for strategy analysis. Visualize equity curves, market patterns, and performance metrics efficiently.
}
\author{
\strong{Maintainer}: Hugo Rzepian Teixeira \email{hugorteixeira@senhormercado.com.br}

}
\keyword{internal}
