% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-utils.R
\name{series_json}
\alias{series_json}
\title{Converts a time series matrix to a JSON-ready list}
\usage{
series_json(mat, datas)
}
\arguments{
\item{mat}{The data matrix.}

\item{datas}{The vector of dates in milliseconds.}
}
\value{
A list ready to be converted to JSON.
}
\description{
Converts a time series matrix to a JSON-ready list
}
\keyword{internal}
