% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-utils.R
\name{stats_json}
\alias{stats_json}
\title{Converts the statistics data.frame to a JSON-ready list}
\usage{
stats_json(carteira_df)
}
\arguments{
\item{carteira_df}{The statistics data.frame.}
}
\value{
A list ready to be converted to JSON.
}
\description{
Converts the statistics data.frame to a JSON-ready list
}
\keyword{internal}
