% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tplot_interactive}
\alias{tplot_interactive}
\title{Launch an Interactive tplot Viewer with Shiny}
\usage{
tplot_interactive(
  ...,
  init = "1994-08-01",
  finit = Sys.Date(),
  rf_rate = NULL,
  geometric = FALSE,
  normalize_risk = NULL,
  modules = c("stats", "candles", "volume", "position", "cumulative", "rolling",
    "period", "drawdowns"),
  theme = default_theme()
)
}
\arguments{
\item{...}{Initial ticker symbols or xts/data.frame objects.}

\item{init}{Start date for the analysis.}

\item{finit}{End date for the analysis.}

\item{rf_rate}{The risk-free rate.}

\item{geometric}{Logical, for geometric returns.}

\item{normalize_risk}{Target annualized volatility (percent) used to rescale series
before plotting. Use `NULL` to keep original risk.}

\item{modules}{Modules to display in the app.}

\item{theme}{The theme function to use.}
}
\value{
Launches a Shiny application. There is no direct return value.
}
\description{
An interactive variant of tplot that opens a Shiny application, allowing the user
to edit, add, and remove assets in real-time to compare performance.
}
