% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-sources.R
\name{.bt_first_text}
\alias{.bt_first_text}
\title{Try to interpret a list as a backtest-like object}
\usage{
.bt_first_text(x)
}
\arguments{
\item{obj}{The list to check.}
}
\value{
A list with mktdata, trades, and stats, or NULL.
}
\description{
Try to interpret a list as a backtest-like object
}
\keyword{internal}
