% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns-risk.R
\name{.normalize_risk}
\alias{.normalize_risk}
\title{Normalize the volatility of a returns series to a target annualized risk in percent.}
\usage{
.normalize_risk(xts, risk = 10, type = c("Discrete", "Log"))
}
\description{
Normalize the volatility of a returns series to a target annualized risk in percent.
}
\keyword{internal}
