% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.tplot_prepare}
\alias{.tplot_prepare}
\title{Prepares all data and metrics for tplot rendering}
\usage{
.tplot_prepare(
  ativo,
  benchs,
  init,
  finit,
  rf_rate,
  auto_rets = FALSE,
  ativo_name = NULL,
  verbose = getOption("tplot.verbose", FALSE)
)
}
\arguments{
\item{ativo}{The main asset (name or xts object).}

\item{benchs}{A list of benchmarks.}

\item{init}{Start date.}

\item{finit}{End date.}

\item{rf_rate}{The risk-free rate.}

\item{auto_rets}{Logical, for geometric returns.}

\item{ativo_name}{An explicit name for the main asset.}
}
\value{
A large list containing all data and metrics prepared for the modules.
}
\description{
Prepares all data and metrics for tplot rendering
}
\keyword{internal}
