% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{rentab_table_calc}
\alias{rentab_table_calc}
\title{Calculates the Monthly and Annual Returns Table}
\usage{
rentab_table_calc(nome_do_objeto, retornar = TRUE, geometric = TRUE)
}
\arguments{
\item{nome_do_objeto}{An xts object with a single column of returns.}

\item{retornar}{Logical, whether to return the table.}

\item{geometric}{Logical, whether to use geometric returns.}
}
\value{
A data.frame with the calendar returns.
}
\description{
Calculates the Monthly and Annual Returns Table
}
\keyword{internal}
