% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{rentab_table_module}
\alias{rentab_table_module}
\title{Renders the Monthly/Annual Returns Table Module}
\usage{
rentab_table_module(returns_tables, asset, benchmarks, theme)
}
\arguments{
\item{returns_tables}{A list of returns data.frames, one for each asset.}

\item{asset}{The name of the main asset.}

\item{benchmarks}{A character vector with benchmark names.}

\item{theme}{The theme list object.}
}
\value{
An HTML object containing the returns tables.
}
\description{
Renders the Monthly/Annual Returns Table Module
}
\keyword{internal}
