% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{stats_module}
\alias{stats_module}
\title{Renders the Performance Statistics Table Module}
\usage{
stats_module(carteira_df, ativo, benchs, theme)
}
\arguments{
\item{carteira_df}{The data.frame with performance statistics.}

\item{ativo}{The name of the main asset.}

\item{benchs}{A character vector with benchmark names.}

\item{theme}{The theme list object.}
}
\value{
An HTML object containing the statistics table.
}
\description{
Renders the Performance Statistics Table Module
}
\keyword{internal}
